12116 op nomina

+++ models/concepto_orden_pago.go
@@ -11,7 +11,7 @@ import (
 
 type ConceptoOrdenPago struct {
 	Id                                               int                                                   `orm:"column(id);pk;auto"`
-	Valor                                            float64                                               `orm:"column(valor)"`
+	Valor                                            int64                                                 `orm:"column(valor)"`
 	Concepto                                         *Concepto                                             `orm:"column(concepto);rel(fk)"`
 	OrdenDePago                                      *OrdenPago                                            `orm:"column(orden_de_pago);rel(fk)"`
 	RegistroPresupuestalDisponibilidadApropiacion    *RegistroPresupuestalDisponibilidadApropiacion        `orm:"column(registro_presupuestal_disponibilidad_apropiacion);rel(fk);null"`
